compute_motor.c: fix warnings unsigned vs signed

+++ C_Controller/src/compute_motor.c
@@ -1,4 +1,5 @@
 #include <stdint.h>
+#include <inttypes.h>
 #include "platform.h"
 #include "recovid.h"
 #include "compute_motor.h"
@@ -15,7 +16,7 @@ unsigned int compute_motor_press_christophe(
 										uint32_t* steps_t_us)
 {
 	unsigned int step_total_us = 0;
-	unsigned int pente_acc, debit_constant, pente_dec;
+	uint32_t pente_acc, debit_constant, pente_dec;
     const uint32_t max_steps = MIN(nb_steps, MAX_MOTOR_STEPS);
 	for(unsigned int i = 0; i < max_steps; i++)
 	{
@@ -62,12 +63,12 @@ unsigned int compute_constant_motor_steps(uint32_t step_t_us, unsigned int nb_st
 		if(i < max_steps) {
 			   int pente_acc = (MOTOR_STEP_TIME_INIT) - ((MOTOR_ACC_COEF)*i);
 				if(pente_acc > 0)
-				   steps_t_us[i] = (uint32_t) MAX(step_t_us, pente_acc);
+				   steps_t_us[i] = (uint32_t) MAX(step_t_us, (uint32_t)pente_acc);
 				else
 				   steps_t_us[i] = step_t_us;
 		}
 		else {
-			   int pente_dec = ((MOTOR_ACC_COEF)* (i-max_steps));
+			   uint32_t pente_dec = ((MOTOR_ACC_COEF)* (i-max_steps));
 			   steps_t_us[i] = MIN(UINT32_MAX,pente_dec);
 		}
 		total_time += step_t_us;
@@ -86,21 +87,21 @@ void print_christophe_header(
 										unsigned int dec_ns,
 										int nb_steps)
 {
-	dbg_printf("step_t_ns_init %d\n", step_t_ns_init);
-	dbg_printf("acc_ns %d\n", acc_ns);
-	dbg_printf("step_t_min_ns %d\n", step_t_min_ns);
-	dbg_printf("speed_down_t_ns %d\n", speed_down_t_ns);
-	dbg_printf("speed_down_t2_ps %d\n", speed_down_t2_ps);
-	dbg_printf("step_t_ns_final %d\n", step_t_ns_final);
-	dbg_printf("dec_ns %d\n", dec_ns);
+	dbg_printf("step_t_ns_init %u\n", step_t_ns_init);
+	dbg_printf("acc_ns %u\n", acc_ns);
+	dbg_printf("step_t_min_ns %u\n", step_t_min_ns);
+	dbg_printf("speed_down_t_ns %u\n", speed_down_t_ns);
+	dbg_printf("speed_down_t2_ps %u\n", speed_down_t2_ps);
+	dbg_printf("step_t_ns_final %u\n", step_t_ns_final);
+	dbg_printf("dec_ns %u\n", dec_ns);
 }
 
 
 void print_steps(uint32_t* steps_t_us, unsigned int nb_steps)
 {
-    printf("Steps %d\n", nb_steps);
+    printf("Steps %u\n", nb_steps);
 	for(unsigned int j=0; j < nb_steps; j+=1)
     {
-		printf("%u\n", steps_t_us[j]);
+		printf("%"PRIu32"\n", steps_t_us[j]);
     }
 }
